backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# YOLO and Computer Vision
ultralytics==8.0.206
opencv-python==4.8.1.78
pillow==10.1.0
numpy<2.0.0

# Real-time features
websockets==12.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=4.0.0
httpx==0.25.2

# Additional dependencies
aiosqlite==0.19.0
shapely==2.0.2
